+++ extensions/meme-generator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Meme generator Changelog
 
+## [Added Memegen API provider] - {PR_MERGE_DATE}
+
+- Refactor Imgflip API provider
+- Add Memegen API provider
+- Update dependencies
+
 ## [✨AI Enhancements 🫧] - 2025-04-29
 
 - Add AI-generated meme tool